Add files via upload
Creates a simple spidering program that will go through Twitter accounts and build a database of them.We will start by retrieving one person’s Twitter friends and their statuses, looping through the list of friends, and adding each of the friends to a database to be
retrieved in the future. After we process one person’s Twitter friends, we check in our database and retrieve one of the friends of the friend. We do this over and over, picking an “unvisited” person, retrieving their friend list, and adding friends
we have not seen to our list for a future visit. We also track how many times we have seen a particular friend in the database to
get some sense of their “popularity”. By storing our list of known accounts and whether we have retrieved the account
or not, and how popular the account is in a database on the disk of the computer, we can stop and restart our program as many times as we like.